Consolidate diamond source into module.py and diamond.py
- Merge 9 source files into two: module.py (models) and diamond.py (agent/logic)
- Delete: unet.py, edm.py, edm_sampling.py, edm_train.py, diffusion.py, agent.py, examples.py, diamond_agent.py, cost_model.py
- Update all imports in tests, scripts, configs, and wm/__init__.py
- All 27 tests pass
