+++ stable_worldmodel/wm/__init__.py
@@ -7,5 +7,3 @@ from .prejepa import *  # noqa: F403
 from .lewm import *  # noqa: F403
 from .pldm import *  # noqa: F403
 from .diamond import *  # noqa: F403
-from .diamond.diffusion import *  # noqa: F403
-from .diamond.edm import *  # noqa: F403
