+++ stable_worldmodel/wm/diamond/__init__.py
@@ -1,8 +1,2 @@
-from .diffusion import *  # noqa: F403
-from .edm import *  # noqa: F403
-from .edm_sampling import *  # noqa: F403
-from .edm_train import *  # noqa: F403
-from .unet import *  # noqa: F403
-from .agent import *  # noqa: F403
-from .diamond_agent import *  # noqa: F403
-from .cost_model import *  # noqa: F403
+from .module import *  # noqa: F403
+from .diamond import *  # noqa: F403
